c_api: Explicitly cast timeval.tv_sec to i64.
On Windows, timeval.tv_sec is an i32, but chrono needs an i64.

+++ src/c_api/connection_parser.rs
@@ -26,7 +26,7 @@ pub unsafe extern "C" fn htp_connp_close(
             timestamp
                 .as_ref()
                 .map(|val| {
-                    NaiveDateTime::from_timestamp_opt(val.tv_sec, val.tv_usec as u32)
+                    NaiveDateTime::from_timestamp_opt(val.tv_sec as i64, val.tv_usec as u32)
                         .map(|time| DateTime::<Utc>::from_utc(time, Utc))
                 })
                 .unwrap_or(None),
@@ -129,7 +129,7 @@ pub unsafe extern "C" fn htp_connp_open(
             timestamp
                 .as_ref()
                 .map(|val| {
-                    NaiveDateTime::from_timestamp_opt(val.tv_sec, val.tv_usec as u32)
+                    NaiveDateTime::from_timestamp_opt(val.tv_sec as i64, val.tv_usec as u32)
                         .map(|time| DateTime::<Utc>::from_utc(time, Utc))
                 })
                 .unwrap_or(None),
@@ -152,7 +152,7 @@ pub unsafe extern "C" fn htp_connp_request_close(
             timestamp
                 .as_ref()
                 .map(|val| {
-                    NaiveDateTime::from_timestamp_opt(val.tv_sec, val.tv_usec as u32)
+                    NaiveDateTime::from_timestamp_opt(val.tv_sec as i64, val.tv_usec as u32)
                         .map(|time| DateTime::<Utc>::from_utc(time, Utc))
                 })
                 .unwrap_or(None),
@@ -182,7 +182,7 @@ pub unsafe extern "C" fn htp_connp_request_data(
                 timestamp
                     .as_ref()
                     .map(|val| {
-                        NaiveDateTime::from_timestamp_opt(val.tv_sec, val.tv_usec as u32)
+                        NaiveDateTime::from_timestamp_opt(val.tv_sec as i64, val.tv_usec as u32)
                             .map(|time| DateTime::<Utc>::from_utc(time, Utc))
                     })
                     .unwrap_or(None),
@@ -212,7 +212,7 @@ pub unsafe extern "C" fn htp_connp_response_data(
                 timestamp
                     .as_ref()
                     .map(|val| {
-                        NaiveDateTime::from_timestamp_opt(val.tv_sec, val.tv_usec as u32)
+                        NaiveDateTime::from_timestamp_opt(val.tv_sec as i64, val.tv_usec as u32)
                             .map(|time| DateTime::<Utc>::from_utc(time, Utc))
                     })
                     .unwrap_or(None),
